Rename info comonent to slide-preview Allow arrow-shortcuts everywhere in the preview Resize the preview
