+++ src/app/demo/bootstrap/bootstrap.module.ts
@@ -3,7 +3,7 @@ import {BootstrapComponent} from './bootstrap.component';
 import {RouterModule} from '@angular/router';
 import {SlidesRoutes} from '../../presentation/slide-routes';
 
-import {ExerciseModule} from '../../exercise/exersice.module';
+import {ExerciseModule} from '../../exercise/exercise.module';
 import {PresentationModule} from '../../presentation/presentation.module';
 import {FeedbackModule} from '../../feedback/feedback.module';
 import {BrowserWindowModule} from '../../browser-window/browser-window.module';
