+++ src/app/demo/component-tree/component-tree.module.ts
@@ -3,7 +3,7 @@ import {ComponentTreeComponent} from './component-tree.component';
 import {RouterModule} from '@angular/router';
 import {SlidesRoutes} from '../../presentation/slide-routes';
 
-import {ExerciseModule} from '../../exercise/exersice.module';
+import {ExerciseModule} from '../../exercise/exercise.module';
 import {PresentationModule} from '../../presentation/presentation.module';
 import {FeedbackModule} from '../../feedback/feedback.module';
 import {BrowserWindowModule} from '../../browser-window/browser-window.module';
