+++ src/app/demo/dependency-injection/dependency-injection.module.ts
@@ -3,7 +3,7 @@ import {DependencyInjectionComponent} from './dependency-injection.component';
 import {RouterModule} from '@angular/router';
 import {SlidesRoutes} from '../../presentation/slide-routes';
 import {FeedbackModule} from '../../feedback/feedback.module';
-import {ExerciseModule} from '../../exercise/exersice.module';
+import {ExerciseModule} from '../../exercise/exercise.module';
 import {PresentationModule} from '../../presentation/presentation.module';
 import {Ng2TsExercises} from '../../../../ng2ts/ng2ts';
