+++ src/app/demo/typescript/typescript.module.ts
@@ -2,7 +2,7 @@ import {NgModule} from '@angular/core';
 import {TypescriptComponent} from './typescript.component';
 import {RouterModule} from '@angular/router';
 import {SlidesRoutes} from '../../presentation/slide-routes';
-import {ExerciseModule} from '../../exercise/exersice.module';
+import {ExerciseModule} from '../../exercise/exercise.module';
 import {PresentationModule} from 'app/presentation/presentation.module';
 import {FeedbackModule} from '../../feedback/feedback.module';
 import {CommonModule} from '@angular/common';
