+++ src/app/exercise/exercise/exercise.component.css
@@ -36,7 +36,7 @@ app-editors {
   height: auto;
 }
 
-app-info {
+app-slides-preview {
   display: flex;
   flex: 1;
 }
