+++ src/app/exercise/exercise/exercise.component.html
@@ -7,8 +7,8 @@
                    (onLoadSolution)="loadSolution($event)">
       </app-editors>
       <app-resize [isVertical]="true">
-        <app-info [milestone]="milestone">
-        </app-info>
+        <app-slides-preview [milestone]="milestone">
+        </app-slides-preview>
       </app-resize>
     </app-resize>
