+++ src/app/exercise/slides-preview/slides-preview.component.html
@@ -11,3 +11,4 @@
     <iframe *ngIf="milestone" [src]="iframeSource"></iframe>
   </div>
 </div>
+
