+++ src/app/exercise/slides-preview/slides-preview.component.scss
@@ -45,7 +45,8 @@
   flex: 1;
   font-weight: bold;
 }
-iframe{
+
+iframe {
   width: 100%;
-  height: 300px;
+  height: 350px;
 }
