+++ src/app/resize/resize.component.ts
@@ -8,7 +8,7 @@ import {Component, ElementRef, HostListener, Input, OnInit} from '@angular/core'
 export class ResizeComponent implements OnInit {
   @Input() isVertical: boolean;
   private MIN_WIDTH = 400;
-  private MIN_HEIGHT = 200;
+  private MIN_HEIGHT = 350;
   private isMouseDown: boolean;
   private initOffset;
   private initSize;
